[Feature] Reason dropdown full scale

+++ MoneyFi-backend/MoneyFi-api-gateway/src/main/java/com/moneyfi/apigateway/model/common/UserAuthHist.java
@@ -18,7 +18,7 @@ public class UserAuthHist {
     private Long id;
     private Long userId;
     private LocalDateTime updatedTime;
-    private Long reasonTypeId;
+    private int reasonTypeId;
     private String comment;
     private Long updatedBy;
 }
